Use Heartbeat
spr currently updates on Stepped. This is done for the unthrottled simulation delta.
Problem: Stepped doesn't run when physics doesn't run. This is a problem for plugin development and apps like Horsecat, which both run in edit mode (no physics). There might be other contexts in the future where it's desirable for spr to update without physics running in the background.
Quick solution for now is to use heartbeat and ignore the simulation delta desync problems. This is acceptable because the use case for simulation time delta is usually in the context of physical or camera animations which use CFrames, and we don't support CFrames yet.

When CFrames are added this will need to switch to an adaptive solution that uses the sim time delta when physics is running, and the wall delta when physics isn't running.