Round-to-nearest UDim offset
UDim[2] Offsets are internally kept as an integer. Offsets constructed from a fractional number are truncated towards zero:
```
print(UDim2.fromOffset(-0.9, 0.9))
--> {0, 0}, {0, 0}
```
This results in a "snapping" behavior, particularly for smaller offset-based position tweens: The sleep limits aren't context aware and fairly tight, so springs go to sleep much later than they appear to stop (usually by a second or two). Until it goes to sleep, the spring is never actually reaching the goal, so a frame is animating from 0->100px, it will appear to rest at 99 for a second until the spring sleeps and it finally hits 100. This transition from 99-100 is usually pretty visible because it's the only motion for a second (Roblox ScrollingFrames have this bug, and you can repro it by scrolling the playerlist).

This fixes the problem by replacing Roblox's towards-zero rounding with nearest-integer rounding.